requirements.txt
ragas>=0.2.0
datasets
openai
anthropic
python-pptx
PyYAML>=6.0.1
